Move DefaultCommandProvider to its own file.
